Modification

+++ src/main/java/com/example/test/AnimalRepository.java
@@ -3,11 +3,14 @@ package com.example.test;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 /*
 Repository for the Animal Table
 
 JpaRepository defines standard CRUD methods, plus JPA-specific operations
  */
 @Repository
-public interface AnimalRepository extends JpaRepository<Animal, Long> {
+public interface AnimalRepository extends JpaRepository<Animal, Integer> {
+
 }
